Expose more hf.co API surface, and ability to download a snapshot of a repo to a local folder (#25)
* [ci] Slim down the matrix quite a bit

* Expose more hf.co API surface

- Expose the `filter` param on the list of models to only list models with a specific tag (method renamed to list_models, don't think it's too breaking, feedback welcome)

- Expose a `model_info` method that gives you slightly more detailed info on a specific model, at a specific revision

* Add `snapshot_download`: Download a whole snapshot of a repo's files at the specified revision

* Docs, tests, lockfiles & metadata

* In `force_filename` mode do not even store metadata file

* Update README.md

+++ src/huggingface_hub/file_download.py
@@ -281,6 +281,7 @@ def cached_download(
     cache_dir: Union[str, Path, None] = None,
     user_agent: Union[Dict, str, None] = None,
     force_download=False,
+    force_filename: Optional[str] = None,
     proxies=None,
     etag_timeout=10,
     resume_download=False,
@@ -360,7 +361,9 @@ def cached_download(
             # etag is None
             pass
 
-    filename = url_to_filename(url, etag)
+    filename = (
+        force_filename if force_filename is not None else url_to_filename(url, etag)
+    )
 
     # get cache path to put the file
     cache_path = os.path.join(cache_dir, filename)
@@ -378,7 +381,11 @@ def cached_download(
                 )
                 if not file.endswith(".json") and not file.endswith(".lock")
             ]
-            if len(matching_files) > 0 and not force_download:
+            if (
+                len(matching_files) > 0
+                and not force_download
+                and force_filename is None
+            ):
                 return os.path.join(cache_dir, matching_files[-1])
             else:
                 # If files cannot be found and local_files_only=True,
@@ -444,10 +451,11 @@ def cached_download(
         logger.info("storing %s in cache at %s", url, cache_path)
         os.replace(temp_file.name, cache_path)
 
-        logger.info("creating metadata file for %s", cache_path)
-        meta = {"url": url, "etag": etag}
-        meta_path = cache_path + ".json"
-        with open(meta_path, "w") as meta_file:
-            json.dump(meta, meta_file)
+        if force_filename is None:
+            logger.info("creating metadata file for %s", cache_path)
+            meta = {"url": url, "etag": etag}
+            meta_path = cache_path + ".json"
+            with open(meta_path, "w") as meta_file:
+                json.dump(meta, meta_file)
 
     return cache_path
